requirements.txt
# Core ML/DL dependencies
tensorflow>=2.13.0,<3.0.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0
pandas>=2.0.0,<3.0.0

# Computer vision
opencv-python>=4.8.0,<5.0.0
Pillow>=10.0.0,<11.0.0

# Visualization
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0
plotly>=5.15.0,<6.0.0

# MLOps and experiment tracking
mlflow>=2.5.0,<3.0.0
wandb>=0.15.0,<1.0.0

# Configuration management
pyyaml>=6.0,<7.0
omegaconf>=2.3.0,<3.0.0

# API and CLI
fastapi>=0.100.0,<1.0.0
uvicorn>=0.22.0,<1.0.0
typer>=0.9.0,<1.0.0
click>=8.1.0,<9.0.0

# Utilities
tqdm>=4.65.0,<5.0.0
python-dotenv>=1.0.0,<2.0.0
pathlib2>=2.3.0,<3.0.0

# Data handling
h5py>=3.9.0,<4.0.0
pyarrow>=12.0.0,<14.0.0

# Development and testing
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0,<5.0.0
black>=23.7.0,<24.0.0
flake8>=6.0.0,<7.0.0
mypy>=1.5.0,<2.0.0
pre-commit>=3.3.0,<4.0.0

# Documentation
sphinx>=7.1.0,<8.0.0
sphinx-rtd-theme>=1.3.0,<2.0.0
mkdocs>=1.5.0,<2.0.0
mkdocs-material>=9.1.0,<10.0.0

# Jupyter notebooks
jupyter>=1.0.0,<2.0.0
jupyterlab>=4.0.0,<5.0.0
ipywidgets>=8.0.0,<9.0.0

# Model serving and deployment  
docker>=6.1.0,<7.0.0
kubernetes>=27.2.0,<28.0.0